Add simple SOM2D demo
SOM2D contains a neural net which trains from an arbitrary text file.

stdout is intended to visually represent the state of the network
topology as it trains on the text file. The <enter> key should train on
the next character then print the network representation again.

The network visualization is intended to demonstrate behavior of things
like neighborhood influence algorithms so I can eventually experiment
with things like wraparound influence (activating a unit in a corner
influences neighbors near all corners).